Moved framework to a Coq plug-in.
Instead of relying on a makefile, running a test can now be done using the
Gallina "QuickCheck" command. For now, the command takes a global definition as
argument (typically c where c := quickCheck p, p being the property to test).

Still experimental, but seems to work. The next steps are:
- A more reliable handling of the name of the property to be tested.
- Take as argument the constr expressing the property directly.
- Make extraction more modular to avoid recompiling a lot of stuff for run of
tests.