Small sets library to express point free set reasoning for generators semantics.
Rewrote most of the proofs in GenLow.
Had to admit some in GenHigh and SemChecker, working on it.

The biggest change I made was that generators are now nat -> RandomSeed -> T
instead of RandomSeed -> nat -> T which makes it easier to use the monad
operators with explicit sizes. It may have broken other things though.