Remove dependency on NPeano.
This file (from Coq stdlib) is deprecated, and it turns out that using
it triggers an extraction bug (the .ml file produced by Separate
Extraction does not match the .mli file).

We should still try to minimize and report the extraction bug, but from
QuickChick's point of view, this is an easy fix.