Tutorial: mention output location within VS Code's VSCoq plugin
Per #267, VSCoq places it in a non-obvious location, so it appears as though
`QuickChick` is a no-op.  While this could be documented in the README or some
other top-level document, I think having it front-and-centre in the tutorial is
optimal as I imagine that's the location where newcomers are most likely to
trip over things.